--- lib/timer.dart
@@ -1,51 +1,0 @@
-import 'package:flutter/material.dart';
-import 'dart:async';
-
-class TimerWidget extends StatefulWidget {
-  const TimerWidget({Key? key}) : super(key: key);
-
-  @override
-  _TimerWidgetState createState() => _TimerWidgetState();
-}
-
-class _TimerWidgetState extends State<TimerWidget> {
-  late Timer _timer;
-  int _secondsElapsed = 0;
-
-  @override
-  void initState() {
-    super.initState();
-    startTimer();
-  }
-
-  void startTimer() {
-    const oneSec = Duration(seconds: 1);
-    _timer = Timer.periodic(oneSec, (Timer timer) {
-      setState(() {
-        _secondsElapsed++;
-      });
-    });
-  }
-
-  String formatTime(int seconds) {
-    int minutes = seconds ~/ 60;
-    int remainingSeconds = seconds % 60;
-    String formattedTime =
-        '${minutes.toString().padLeft(2, '0')}:${remainingSeconds.toString().padLeft(2, '0')}';
-    return formattedTime;
-  }
-
-  @override
-  void dispose() {
-    _timer.cancel();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Text(
-      formatTime(_secondsElapsed),
-      style: TextStyle(fontSize: 30, fontWeight: FontWeight.bold),
-    );
-  }
-}
